refactor: 하나의 클래스로 재통합
