+++ backend/src/main/java/hyfive/gachita/api/kakao/KakaoNaviApiClient.java
@@ -15,22 +15,39 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
 
-// 다중 경유지 길찾기 API : 최대 30개의 경유지 추가 가능
+// 자동차 길찾기 API : 최대 5개의 경유지 추가 가능
 @Component
 @Slf4j
-public class KakaoWaypointsApiClient extends ApiClient {
+public class KakaoNaviApiClient extends ApiClient {
 
+    private static final String DIRECTIONS_URI = "/directions";
     private static final String WAYPOINTS_URI = "/waypoints/directions";
 
     private final String apiKey;
     private final String baseUrl;
 
-    public KakaoWaypointsApiClient(@Qualifier("KakaoNaviRestClient") RestClient restClient,
-                                   @Value("${kakao.navi.api.key}") String apiKey,
-                                   @Value("${kakao.navi.api.base-url}") String baseUrl) {
+    public KakaoNaviApiClient(@Qualifier("KakaoNaviRestClient") RestClient restClient,
+                                    @Value("${kakao.navi.api.key}") String apiKey,
+                                    @Value("${kakao.navi.api.base-url}") String baseUrl) {
         super(restClient);
         this.apiKey = apiKey;
         this.baseUrl = baseUrl;
+    }
+
+    public KakaoNaviRes.Route getDirections(DirectionsReq request) {
+        URI uri = UriComponentsBuilder.fromUriString(baseUrl + DIRECTIONS_URI)
+                .queryParam("origin", request.origin())
+                .queryParam("destination", request.destination())
+                .build()
+                .toUri();
+
+        KakaoNaviRes result = restClient.get()
+                .uri(uri)
+                .header("Authorization", "KakaoAK " + apiKey)
+                .retrieve()
+                .body(KakaoNaviRes.class);
+
+        return getValidRoute(result);
     }
 
     public KakaoNaviRes.Route getWaypointsDirections(DirectionsReq request) {
@@ -43,6 +60,10 @@ public class KakaoWaypointsApiClient extends ApiClient {
                 .retrieve()
                 .body(KakaoNaviRes.class);
 
+        return getValidRoute(result);
+    }
+
+    private static KakaoNaviRes.Route getValidRoute(KakaoNaviRes result) {
         KakaoNaviRes.Route route = result.routes().get(0);  // 첫 번째 추천 경로 기준
         if (route.resultCode() != NaviResultCode.SUCCSS.code) {
             throw new BusinessException(ErrorCode.EXTERNAL_API_ERROR);
